+++ src/Job/JobSchedule.php
@@ -4,6 +4,7 @@ namespace Orisai\Scheduler\Job;
 
 use Closure;
 use Cron\CronExpression;
+use DateTimeZone;
 
 final class JobSchedule
 {
@@ -18,13 +19,16 @@ final class JobSchedule
 	/** @var int<0, 30> */
 	private int $repeatAfterSeconds;
 
+	private ?DateTimeZone $timeZone;
+
 	/**
 	 * @param int<0, 30> $repeatAfterSeconds
 	 */
 	public static function create(
 		Job $job,
 		CronExpression $expression,
-		int $repeatAfterSeconds
+		int $repeatAfterSeconds,
+		?DateTimeZone $timeZone = null
 	): self
 	{
 		$self = new self();
@@ -32,6 +36,7 @@ final class JobSchedule
 		$self->job = $job;
 		$self->expression = clone $expression;
 		$self->repeatAfterSeconds = $repeatAfterSeconds;
+		$self->timeZone = $timeZone;
 
 		return $self;
 	}
@@ -43,13 +48,15 @@ final class JobSchedule
 	public static function createLazy(
 		Closure $jobConstructor,
 		CronExpression $expression,
-		int $repeatAfterSeconds
+		int $repeatAfterSeconds,
+		?DateTimeZone $timeZone = null
 	): self
 	{
 		$self = new self();
 		$self->jobConstructor = $jobConstructor;
 		$self->expression = clone $expression;
 		$self->repeatAfterSeconds = $repeatAfterSeconds;
+		$self->timeZone = $timeZone;
 
 		return $self;
 	}
@@ -75,6 +82,11 @@ final class JobSchedule
 	public function getRepeatAfterSeconds(): int
 	{
 		return $this->repeatAfterSeconds;
+	}
+
+	public function getTimeZone(): ?DateTimeZone
+	{
+		return $this->timeZone;
 	}
 
 }